food/pyproject.toml
[tool.poetry]
name = "food"
version = "0.1.0"
description = ""
authors = ["ig-rudenko <ig.rudenko1@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
pillow = "^10.2.0"
django-ckeditor = "^6.7.0"
django-crispy-forms = "^2.1"
crispy-bootstrap5 = "^2023.10"
psycopg2 = "^2.9.9"
djangorestframework = "^3.14.0"
django-filter = "^23.5"
djoser = "^2.2.2"
requests = "^2.31.0"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.3.1"}
django-debug-toolbar = "^4.2.0"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
